Version 1.13 R2
Started working on NMS 1.13_R2.

I'm needing to find a stable method of keeping the NPC in a sleeping position and keeping the skin active. This is because the 'Pose' Field on an EntityPlayer was added in 1.14 and so I need to create a fake bed in 1.13 and below to make the corpse functional..